assigment completed

+++ programs/er-state-account/src/state/user_account.rs
@@ -3,7 +3,7 @@ use anchor_lang::prelude::*;
 #[account]
 pub struct UserAccount {
     pub user: Pubkey,
-    pub data: u64,
+    pub data: u8, 
     pub bump: u8,
 }
